+++ tests/test_lecsem.py
@@ -11,10 +11,10 @@ XFILE, VAR = './tests/data/chasim_hallue_2var.out', "CHARGE"
 def test_lecsem_scan_dim():
 
     nu_zoomx = core.modgridmarthe.scan_nu_zoomx(XFILE) # scan nb of nested grids (gig)
-    dims, nbsteps = core.modgridmarthe.scan_dim(XFILE, VAR, nu_zoomx)
+    dims, nbsteps = core.modgridmarthe.scan_dim(XFILE, VAR)  # nu_zoomx
 
     assert isinstance(dims, np.ndarray)
-    assert dims.shape[0] == nu_zoomx + 1
+    # assert dims.shape[0] == nu_zoomx + 1
     assert dims.shape[1] == 3  # x, y, z
     assert isinstance(nbsteps, int)
     assert nbsteps > 0